refactor(user): move files into domain-based folders, update imports, and remove unused controller test
- Created folders: controller/, service/, repository/
- Moved files into corresponding folders
- Updated import paths in UserModule and related test files
- Remove unused user.controller.spec.ts (no longer relevant to current controller logic)
- Verified no business logic changes, all imports resolved successfully
